Alternative diagram for initial/terminal objects
In the original diagram, all objects are both initial and terminal. I think the idea was to show that terminal objects can still have outgoing arrows, and initial objects can still have incoming arrows. But maybe it would be clearer to have a diagram where the initial and terminal objects aren't isomorphic.

On the other hand, now that I've drawn this out, I'm not sure it is much of an improvement!